% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_scan_pvl.R
\name{add_phenames}
\alias{add_phenames}
\title{Replace old (default) trait names with true trait names}
\usage{
add_phenames(charvec, phe1, phe2)
}
\arguments{
\item{charvec}{a character vector containing entries like "profile1" and "profile2"}

\item{phe1}{character vector of length 1 that is the true trait name for the first trait}

\item{phe2}{character vector of length 1 that is the true trait name for the second trait}
}
\description{
Replace old (default) trait names with true trait names
}
