% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_scan_pvl.R
\name{add_pmap}
\alias{add_pmap}
\title{Add physical map contents to tibble}
\usage{
add_pmap(tib, pmap)
}
\arguments{
\item{tib}{a tibble with 3 columns: marker1 name, marker2 name and log-likelihood values}

\item{pmap}{a physical map for a single chromosome}
}
\description{
Add physical map contents to tibble
}
