% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dimnames.R
\name{arrange_by_rownames}
\alias{arrange_by_rownames}
\title{Arrange rows of a matrix or data frame so that its rownames match those of another matrix or data frame}
\usage{
arrange_by_rownames(df, ref)
}
\arguments{
\item{df}{the object whose rows will be reordered}

\item{ref}{the object whose rows will not be reordered}
}
\value{
df, but with rows in different order, so that the rownames match those of ref
}
\description{
Arrange rows of a matrix or data frame so that its rownames match those of another matrix or data frame
}
