% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_scan_pvl.R
\name{assemble_profile_tib}
\alias{assemble_profile_tib}
\title{Assemble a profile ll tibble}
\usage{
assemble_profile_tib(tib, trace = "profile1")
}
\arguments{
\item{tib}{a tibble, derived from loglik_mat and with pmap info}

\item{trace}{character vector to identify which profile log likelihood to calculate. Takes values "profile1" or "profile2"}
}
\description{
Assemble a profile ll tibble
}
