% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_n.R
\name{boot_n}
\alias{boot_n}
\title{Create a bootstrap sample and perform multivariate QTL scan}
\usage{
boot_n(pp, pleio_peak_index, phe, covariates = NULL, kinship,
  nboot_per_job = 1, start_snp1, n_snp)
}
\arguments{
\item{pp}{allele probabilities object for one chromosome only, like aprobs$`8`. Not a list}

\item{pleio_peak_index}{positive integer index indicating design matrix for simulation}

\item{phe}{n by d matrix of phenotypes}

\item{covariates}{n by n.cov matrix of covariates}

\item{kinship}{a kinship MATRIX, not a list}

\item{nboot_per_job}{number of bootstrap samples to call per invocation of function}

\item{start_snp1}{positive integer indicating index within pp for start of scan}

\item{n_snp}{number of markers to use in scan}
}
\value{
numerical vector of lrt statistics from one or more bootstrap samples
}
\description{
Create a bootstrap sample and perform multivariate QTL scan
}
