% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_pvl.R
\name{boot_pvl}
\alias{boot_pvl}
\title{Create a bootstrap sample and perform multivariate QTL scan}
\usage{
boot_pvl(probs, pleio_peak_index, pheno, covariates = NULL,
  kinship = NULL, nboot_per_job = 1, start_snp, n_snp)
}
\arguments{
\item{probs}{allele probabilities object for one chromosome only, like aprobs$`8`. Not a list}

\item{pleio_peak_index}{positive integer index indicating design matrix for simulation. Typically acquired by using `find_pleio_peak_tib`.}

\item{pheno}{n by d matrix of phenotypes}

\item{covariates}{n by n.cov matrix of covariates}

\item{kinship}{a kinship matrix, not a list}

\item{nboot_per_job}{number of bootstrap samples to call per invocation of function}

\item{start_snp}{positive integer indicating index within probs for start of scan}

\item{n_snp}{number of (consecutive) markers to use in scan}
}
\value{
numerical vector of lrt statistics from `nboot_per_job` bootstrap samples
}
\description{
Create a bootstrap sample and perform multivariate QTL scan
}
