% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_Bhat.R
\name{calc_Bhat}
\alias{calc_Bhat}
\title{Calculate estimated allele effects, B matrix}
\usage{
calc_Bhat(X, Sigma, Y)
}
\arguments{
\item{X}{2n by 2f block-diagonal design matrix that incorporates genetic info for two markers. Note that we can use the same marker data twice.}

\item{Sigma}{2n by 2n covariance matrix, often composed as \eqn{K \otimes V_g + I_n \otimes V_e}}

\item{Y}{2n by 1 matrix, ie, a column vector, of 2 phenotypes' measurements}
}
\description{
Calculate estimated allele effects, B matrix
}
