% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_sqrt_mat.R
\name{calc_invsqrt_mat}
\alias{calc_invsqrt_mat}
\title{Calculate matrix inverse square root for a covariance matrix}
\usage{
calc_invsqrt_mat(A)
}
\arguments{
\item{A}{covariance matrix}
}
\description{
Calculate matrix inverse square root for a covariance matrix
}
