% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_loglik_bvlmm.R
\name{calc_loglik_bvlmm}
\alias{calc_loglik_bvlmm}
\title{Calculate log likelihood for a bivariate linear mixed effects model}
\usage{
calc_loglik_bvlmm(X, Bhat, Y, Sigma)
}
\arguments{
\item{X}{a block-diagonal, 2n by 2f design matrix containing genotype probabilities for two markers}

\item{Bhat}{a vector of estimated allele effects. Can be generated by as.vector(B), where B is f by 2}

\item{Y}{a n by 2 matrix of phenotype values}

\item{Sigma}{a 2n by 2n covariance matrix.}
}
\description{
Calculate log likelihood for a bivariate linear mixed effects model
}
