% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_lrt.R
\name{calc_lrt-deprecated}
\alias{calc_lrt-deprecated}
\title{Calculate the LRT statistic}
\usage{
calc_lrt(loglik_mat)
}
\arguments{
\item{loglik_mat}{a matrix of log likelihood values outputted by scan_pvl()}
}
\description{
Calculate the LRT statistic
}
\seealso{
\code{\link{qtl2pleio-deprecated}}
}
\keyword{internal}
