% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_lrt_tib.R
\name{calc_lrt_tib}
\alias{calc_lrt_tib}
\title{Calculate a LRT statistic from the (new) output from scan_pvl}
\usage{
calc_lrt_tib(tib)
}
\value{
a number, the LRT statistic
}
\description{
Calculate a LRT statistic from the (new) output from scan_pvl
}
