% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_profile_lods.R
\name{calc_profile_lods}
\alias{calc_profile_lods}
\title{Calculate profile lods for all traits}
\usage{
calc_profile_lods(scan_pvl_out)
}
\arguments{
\item{scan_pvl_out}{tibble outputted from scan_pvl}
}
\description{
Calculate profile lods for all traits
}
