% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_Sigma.R
\name{calc_Sigma}
\alias{calc_Sigma}
\title{Calculate the phenotypes covariance matrix Sigma}
\usage{
calc_Sigma(Vg, Ve, K)
}
\arguments{
\item{Vg}{2 by 2 genetic covariance matrix for the two phenotypes}

\item{Ve}{2 by 2 error covariance matrix for the two phenotypes}

\item{K}{n by n kinship matrix}
}
\description{
Calculate the phenotypes covariance matrix Sigma
}
