% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_sqrt_mat.R
\name{calc_sqrt_mat}
\alias{calc_sqrt_mat}
\title{Calculate matrix square root for a covariance matrix}
\usage{
calc_sqrt_mat(A)
}
\arguments{
\item{A}{covariance matrix}
}
\description{
Calculate matrix square root for a covariance matrix
}
