% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dimnames.R
\name{check_dimnames}
\alias{check_dimnames}
\title{Verify that a single dimension's dimnames of two objects are equal}
\usage{
check_dimnames(x, y, xdim = 1, ydim = 1)
}
\arguments{
\item{x}{first object}

\item{y}{second object}

\item{xdim}{dimension to examine for first object}

\item{ydim}{dimension to examine for second object}
}
\value{
logical indicator of equality of two objects' rownames
}
\description{
Verify that a single dimension's dimnames of two objects are equal
}
