% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_missingness}
\alias{check_missingness}
\title{Check for missingness in phenotypes or covariates}
\usage{
check_missingness(input_matrix)
}
\arguments{
\item{input_matrix}{phenotypes or covariates matrix}
}
\value{
logical vector indicating presence of no missingness. Those entries with TRUE have no missingness, while those with FALSE have one or more missing values.
}
\description{
Check for missingness in phenotypes or covariates
}
