% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_missing_jobs.R
\name{find_missing_jobs}
\alias{find_missing_jobs}
\title{Find jobs ids of missing files from CHTC}
\usage{
find_missing_jobs(DIR = "results-chtc/boot400-run1", njobs = 16000,
  outfile = "bad-jobs-boot400-run1.txt")
}
\arguments{
\item{DIR}{directory for results files}

\item{njobs}{number of jobs expected - typically 400 or 16000}

\item{outfile}{file path for writing missing job ids}
}
\description{
Find jobs ids of missing files from CHTC
}
