% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_li_cols.R
\name{get_li_cols}
\alias{get_li_cols}
\title{Extract a subset of covariate columns that are linearly independent}
\usage{
get_li_cols(addcovar, add_intercept = TRUE)
}
\arguments{
\item{addcovar}{matrix containing additive covariates}

\item{add_intercept}{logical indicating whether to add an intercept column when checking linear independence.}
}
\description{
We have the option of adding or not adding an intercept before doing the QR decomposition,
as indicated by the input `add_intercept`.
We will typically use `add_intercept = TRUE` because, in practice,
`find_li_cols` is called within `scan_pvl`, the function that
performs the two-dimensional scan. Within `scan_pvl`, we form a design matrix
that consists of a founder allele probabilities matrix AND the covariates.
To ensure that this matrix is invertible, we need to consider the possibility
appending the covariates to the founder allele probabilities matrix leads
to a matrix that is not full rank, possibly due to the fact that
the founder allele probabilities add to 1 and that a subset of the covariate columns
may sum to 1 (or a multiple of 1).
}
\examples{
c1 <- rep(c(0, 1), each = 4)
c2 <- rep(c(0, 1), times = 4)
c3 <- rep(c(0, 0, 1, 1), times = 2)
addcovar <- cbind(c1, c2, c3)
}
\references{
https://stackoverflow.com/questions/14943422/r-independent-columns-in-matrix
}
