% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_recombination.R
\name{id_recombination}
\alias{id_recombination}
\title{Determine if all genotype probability vectors for markers in an interval are the same for one subject.}
\usage{
id_recombination(probs)
}
\arguments{
\item{probs}{a genotype probabilities 2-dimensional array for the appropriate chromosome and a single subject}
}
\description{
Determine if all genotype probability vectors for markers in an interval are the same for one subject.
}
\details{
Checks genotype probability vector (at each marker in probs) for equality with the left-hand marker's genotype probabilities vector.
}
