% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_id2keep}
\alias{make_id2keep}
\title{Identify shared subject ids among all inputs: covariates, allele probabilities array, kinship, and phenotypes}
\usage{
make_id2keep(probs, pheno, covar, kinship)
}
\arguments{
\item{probs}{an allele probabilities array}

\item{pheno}{a phenotypes matrix}

\item{covar}{a covariates matrix}

\item{kinship}{a kinship matrix}
}
\description{
Identify shared subject ids among all inputs: covariates, allele probabilities array, kinship, and phenotypes
}
