% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_results_table.R
\name{make_results_table}
\alias{make_results_table}
\title{Read boot output files - which contain lrt statistics - and assemble into a single table}
\usage{
make_results_table(DIR = "results-chtc/boot400-run1")
}
\arguments{
\item{DIR}{directory path for directory that contains boot output files}
}
\description{
Read boot output files - which contain lrt statistics - and assemble into a single table
}
