% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pvl.R
\name{plot_pvl}
\alias{plot_pvl}
\title{Plot tidied results of a pvl scan}
\usage{
plot_pvl(dat, phenames, size = 3, shape = 17, palette = c("#999999",
  "#E69F00", "#56B4E9", "#009E73", "#F0E442", "#0072B2", "#D55E00", "#CC79A7"))
}
\arguments{
\item{dat}{a tibble that is the tidied output of a pvl scan}

\item{phenames}{vector of names of the two ordered phenotypes}

\item{palette}{a vector containing strings for colors}
}
\description{
Plot tidied results of a pvl scan
}
