% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_mytab.R
\name{prep_X_list}
\alias{prep_X_list}
\title{Create a list of component X matrices for input to stagger_mats, to ultimately create design matrix}
\usage{
prep_X_list(indices, start_snp, probs, covariates)
}
\arguments{
\item{indices}{a vector of integers}

\item{start_snp}{an integer}

\item{probs}{a three-dimensional array of genotype probabilities for a single chromosome}

\item{covariates}{a matrix of covariates}
}
\description{
Create a list of component X matrices for input to stagger_mats, to ultimately create design matrix
}
