% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_recombination.R
\name{probs_equal}
\alias{probs_equal}
\title{Determine if a single subject's genotype probabilities at two markers are equal.}
\usage{
probs_equal(g1, g2)
}
\arguments{
\item{g1}{a vector of genotype probabilities for a single subject at marker one}

\item{g2}{a vector of genotype probabilities for the same subject at marker two}
}
\description{
Determine if a single subject's genotype probabilities at two markers are equal.
}
