% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_inputs.R
\name{process_inputs}
\alias{process_inputs}
\title{Process inputs}
\usage{
process_inputs(probs, pheno, addcovar, kinship)
}
\arguments{
\item{probs}{a three-dimensional array of founder allele probabilities}

\item{pheno}{a matrix of d trait values}

\item{addcovar}{a matrix of covariates}

\item{kinship}{a kinship matrix}
}
\description{
Process inputs
}
