% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvl_boot.R
\name{pvl_boot}
\alias{pvl_boot}
\title{Calculate likelihood ratio test statistics for simulated data sets for a single marker}
\usage{
pvl_boot(X, B, Vg_initial, Ve_initial, kinship, probs, start_snp, n_snp,
  nboot = 1000)
}
\arguments{
\item{X}{a genotypes probability matrix for a single marker, ie, the marker that has the peak LOD under pleiotropy}

\item{B}{allele effects matrix, often calculated via calc_Bhat}

\item{Vg_initial}{estimated Vg covariance matrix from true data}

\item{Ve_initial}{estimated Ve covariance matrix from true data}

\item{kinship}{a chromosome-specific kinship matrix}

\item{probs}{a single chromosome's genotype probabilities}

\item{start_snp}{index for starting the scan}

\item{n_snp}{number of SNPs to include in the scan}

\item{nboot}{number of bootstrap samples}
}
\description{
Calculate likelihood ratio test statistics for simulated data sets for a single marker
}
