% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl2pleio-package.r
\docType{package}
\name{qtl2pleio-package}
\alias{qtl2pleio}
\alias{qtl2pleio-package}
\title{qtl2pleio.}
\description{
Testing pleiotropy vs. separate QTL in multiparental populations
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fboehm/qtl2pleio}
  \item Report bugs at \url{https://github.com/fboehm/qtl2pleio/issues}
}

}
\author{
\strong{Maintainer}: Frederick J Boehm \email{frederick.boehm@gmail.com} (\href{https://orcid.org/0000-0002-1644-5931}{ORCID})

}
