% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcpp_calc_Bhat.R
\name{rcpp_calc_Bhat}
\alias{rcpp_calc_Bhat}
\title{Estimate allele effects matrix, B hat, with Rcpp}
\usage{
rcpp_calc_Bhat(X, Sigma_inv, Y)
}
\arguments{
\item{X}{dn by df block-diagonal design matrix that incorporates genetic info for two markers. Note that we can use the same marker data twice.}

\item{Sigma_inv}{dn by dn inverse covariance matrix, often composed as the inverse of \eqn{K \otimes V_g + I_n \otimes V_e}}

\item{Y}{dn by 1 matrix, ie, a column vector, of d phenotypes' measurements}
}
\value{
a df by 1 matrix of GLS-estimated allele effects
}
\description{
Estimate allele effects matrix, B hat, with Rcpp
}
