% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcpp_log_dmvnorm2.R
\name{rcpp_log_dmvnorm2}
\alias{rcpp_log_dmvnorm2}
\title{Calculate log likelihood for a multivariate normal}
\usage{
rcpp_log_dmvnorm2(inv_S, mu, x, S)
}
\arguments{
\item{inv_S}{inverse covariance matrix}

\item{mu}{mean vector}

\item{x}{data vector}

\item{S}{covariance matrix, ie, the inverse of inv_S}
}
\description{
Calculate log likelihood for a multivariate normal
}
