% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_boot_lrt.R
\name{read_boot_lrt}
\alias{read_boot_lrt}
\title{Read LRT statistics for all of the scans (ie, 400 typically) that correspond to a single simulated trait file}
\usage{
read_boot_lrt(job_num = 0, total_number_of_scans = 400,
  nlrt_per_file = 10, boot_run_num = 1)
}
\arguments{
\item{job_num}{trait file id number. An integer in 0:399 typically}

\item{total_number_of_scans}{number of scans ie number of 2d scans - for each trait file}

\item{nlrt_per_file}{positive integer specifying number of lrt statistics per boot output file}

\item{boot_run_num}{run number for the boot files. Used to create the directory path and file names}
}
\description{
Read LRT statistics for all of the scans (ie, 400 typically) that correspond to a single simulated trait file
}
