% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_pvl.R
\name{scan_pvl}
\alias{scan_pvl}
\title{Perform model fitting for all ordered pairs of markers in a genomic region of interest}
\usage{
scan_pvl(probs, pheno, kinship, start_snp1, start_snp2 = start_snp1, n_snp)
}
\arguments{
\item{probs}{an array of genotype probabilities for a single chromosome (before eigenrotation)}

\item{pheno}{a matrix of phenotypes (before eigenrotation)}

\item{kinship}{a kinship matrix for one chromosome}

\item{start_snp1}{index of where to start the first coordinate of the ordered pair}

\item{start_snp2}{index of where to start the second coordinate of the ordered pair. Typically the same value as start_snp1}

\item{n_snp}{the number of (consecutive) snps to include in the scan}
}
\description{
Perform model fitting for all ordered pairs of markers in a genomic region of interest
}
