% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim1.R
\name{sim1}
\alias{sim1}
\title{Simulate a single data set consisting of n subjects and 2 phenotypes for each}
\usage{
sim1(X, B, Vg, Ve, kinship)
}
\arguments{
\item{X}{genotype probabilities matrix for a single marker}

\item{B}{a matrix of allele effects}

\item{Vg}{a genetic covariance matrix}

\item{Ve}{an error covariance matrix}

\item{kinship}{a chromosome-specific kinship matrix}
}
\description{
Simulate a single data set consisting of n subjects and 2 phenotypes for each
}
