% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subset_input}
\alias{subset_input}
\title{Subset an input object - allele probabilities array or phenotypes matrix or covariates matrix. Kinship has its own subset function}
\usage{
subset_input(input, id2keep)
}
\arguments{
\item{input}{a matrix of either phenotypes or covariates or array of allele probabilities}

\item{id2keep}{a character vector of subject ids to identify those subjects that are shared by all inputs}
}
\description{
Subset an input object - allele probabilities array or phenotypes matrix or covariates matrix. Kinship has its own subset function
}
