% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subset_kinship}
\alias{subset_kinship}
\title{Subset a kinship matrix to include only those subjects present in all inputs}
\usage{
subset_kinship(kinship, id2keep)
}
\arguments{
\item{kinship}{a kinship matrix}

\item{id2keep}{a character vector of subject ids to identify those subjects that are shared by all inputs}
}
\description{
Subset a kinship matrix to include only those subjects present in all inputs
}
