% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_pleio_peaks_file.R
\name{write_pleio_peaks_file}
\alias{write_pleio_peaks_file}
\title{Create a pleiotropy peaks file with the marker indices - that designate the pleiotropy peak position - for each of a collection of trait files}
\usage{
write_pleio_peaks_file(run_num, map = pm,
  DIR = paste0("results-chtc/pvl400-run", run_num))
}
\arguments{
\item{run_num}{run number, a positive integer}

\item{map}{a map for the corresponding chromosome}

\item{DIR}{a directory path for finding the 2d scan results}
}
\description{
Create a pleiotropy peaks file with the marker indices - that designate the pleiotropy peak position - for each of a collection of trait files
}
