% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_scan_pvl.R
\name{add_intercepts}
\alias{add_intercepts}
\title{Add intercepts to tidied loglikelihood tibble}
\usage{
add_intercepts(dat, intercepts_univariate)
}
\arguments{
\item{dat}{a tibble that results from tidy_scan_pvl acting on a log likelihood matrix}

\item{intercepts_univariate}{a vector of length 2 that contains the x coordinate values for the ordered univariate peaks}
}
\description{
Add intercepts to tidied loglikelihood tibble
}
\seealso{
Other profile log-likelihood tibble functions: \code{\link{add_pmap}},
  \code{\link{assemble_profile_tib}},
  \code{\link{tidy_scan_pvl}},
  \code{\link{transform_loglik_mat}}
}
\concept{profile log-likelihood tibble functions}
