% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_scan_pvl.R
\name{add_pmap}
\alias{add_pmap}
\title{Add physical map contents to tibble}
\usage{
add_pmap(tib, pmap)
}
\arguments{
\item{tib}{a tibble with 3 columns: marker1 name, marker2 name and log-likelihood values}

\item{pmap}{a physical map for a single chromosome}
}
\value{
a tibble with 5 columns: marker1, marker2, log-likelihood, marker1_position, marker2_position
}
\description{
Add physical map contents to tibble
}
\examples{
pm <- 1:3
names(pm) <- as.character(paste0('m', 1:3))
expand.grid(paste0('m', 1:3), paste0('m', 1:3)) -> foo
tib <- tibble::tibble(marker1 = as.character(foo[ , 1]),
  marker2 = as.character(foo[ , 2]))
tib$ll <- rgamma(9, 5)
add_pmap(tib, pm)
}
\seealso{
Other profile log-likelihood tibble functions: \code{\link{add_intercepts}},
  \code{\link{assemble_profile_tib}},
  \code{\link{tidy_scan_pvl}},
  \code{\link{transform_loglik_mat}}
}
\concept{profile log-likelihood tibble functions}
