% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_scan_pvl.R
\name{assemble_profile_tib}
\alias{assemble_profile_tib}
\title{Assemble a profile log-likelihood tibble}
\usage{
assemble_profile_tib(tib, trace = "profile1")
}
\arguments{
\item{tib}{a tibble, derived from loglik_mat and with pmap info}

\item{trace}{character vector to identify which profile log likelihood to calculate. Takes values 'profile1' or 'profile2'}
}
\value{
a tibble with 3 columns: marker position, profile log-likelihood, and trace (profile1 or profile2)
}
\description{
Assemble a profile log-likelihood tibble
}
\examples{
pm <- 1:3
names(pm) <- as.character(paste0('m', 1:3))
expand.grid(paste0('m', 1:3), paste0('m', 1:3)) -> foo
tib <- tibble::tibble(marker1 = as.character(foo[ , 1]),
  marker2 = as.character(foo[ , 2]))
tib$ll <- rgamma(9, 5)
add_pmap(tib, pm) -> tib_pm
assemble_profile_tib(tib = tib_pm, trace = "profile1")
}
\seealso{
Other profile log-likelihood tibble functions: \code{\link{add_intercepts}},
  \code{\link{add_pmap}}, \code{\link{tidy_scan_pvl}},
  \code{\link{transform_loglik_mat}}
}
