% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_covs.R
\name{calc_covs}
\alias{calc_covs}
\title{Calculate Vg and Ve from Y and kinship}
\usage{
calc_covs(pheno, kinship)
}
\arguments{
\item{pheno}{n by 2 matrix of phenotypes}

\item{kinship}{a kinship matrix, n by n}
}
\description{
Calculate Vg and Ve from Y and kinship
}
