% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_Sigma.R
\name{calc_Sigma}
\alias{calc_Sigma}
\title{Calculate the phenotypes covariance matrix Sigma}
\usage{
calc_Sigma(Vg, Ve, kinship = NULL)
}
\arguments{
\item{Vg}{d by d genetic covariance matrix for the d phenotypes}

\item{Ve}{d by d error covariance matrix for the d phenotypes}

\item{kinship}{optional n by n kinship matrix. if NULL, Vg is not used.}
}
\value{
dn by dn covariance matrix
}
\description{
Calculate the phenotypes covariance matrix Sigma
}
