% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_missingness}
\alias{check_missingness}
\title{Check for missingness in phenotypes or covariates}
\usage{
check_missingness(input_matrix)
}
\arguments{
\item{input_matrix}{phenotypes or covariates matrix}
}
\value{
character vector of subjects that have no missingness
}
\description{
Check for missingness in phenotypes or covariates
}
