% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_recombination.R
\name{id_recombination}
\alias{id_recombination}
\title{Determine if a single subject's lineage had a recombination (historically) between two markers}
\usage{
id_recombination(g1, g2)
}
\arguments{
\item{g1}{a vector of genotype probabilities for a single subject at marker one}

\item{g2}{a vector of genotype probabilities for the same subject at marker two}
}
\description{
Determine if a single subject's lineage had a recombination (historically) between two markers
}
