% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_id2keep}
\alias{make_id2keep}
\title{Identify shared subject ids among all inputs: covariates, allele probabilities array, kinship, and phenotypes}
\usage{
make_id2keep(probs, pheno, addcovar = NULL, kinship = NULL)
}
\arguments{
\item{probs}{an allele probabilities array}

\item{pheno}{a phenotypes matrix}

\item{addcovar}{a covariates matrix}

\item{kinship}{a kinship matrix}
}
\value{
a character vector of subject IDs common to all (non-null) inputs
}
\description{
We consider only those inputs that are not NULL. We then use `intersect` on pairs
of inputs' rownames to identify those subjects are shared among all non-NULL inputs.
}
