% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pvl.R
\name{plot_pvl}
\alias{plot_pvl}
\title{Plot tidied results of a pvl scan}
\usage{
plot_pvl(dat)
}
\arguments{
\item{dat}{a tibble that is the tidied output of a pvl scan}
}
\description{
Plot tidied results of a pvl scan
}
