% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pvl.R
\name{plot_pvl}
\alias{plot_pvl}
\title{Plot tidied results of a pvl scan}
\usage{
plot_pvl(dat, units = "Mb", palette = c("#999999", "#E69F00",
  "#56B4E9"), linetype = c("solid", "longdash", "dotted"))
}
\arguments{
\item{dat}{a profile lod tibble}

\item{units}{a character vector of length one to indicate units for physical or genetic map}

\item{palette}{a character vector of length 3 containing strings for colors}

\item{linetype}{a character vector of length 3 specifying the linetype values for the 3 traces}
}
\value{
a ggplot object with profile LODs
}
\description{
Plot tidied results of a pvl scan
}
