% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_pvl.R
\name{scan_pvl}
\alias{scan_pvl}
\title{Perform model fitting for all ordered pairs of markers in a genomic region of interest}
\usage{
scan_pvl(probs, pheno, kinship, covariates = NULL, start_snp1 = 1, n_snp,
  max_iter = 1e+05, max_prec = 1/1e+06)
}
\arguments{
\item{probs}{an array of genotype probabilities for a single chromosome (before eigenrotation)}

\item{pheno}{a matrix of phenotypes (before eigenrotation)}

\item{kinship}{a kinship matrix for one chromosome}

\item{covariates}{a matrix, n subjects by n.cov covariates, where each column is one covariate}

\item{start_snp1}{index of where to start the scan within probs}

\item{n_snp}{the number of (consecutive) snps to include in the scan}

\item{max_iter}{maximum number of iterations for EM algorithm}

\item{max_prec}{stepwise precision for EM algorithm. EM stops once incremental difference in log likelihood is less than max_prec}
}
\value{
a tibble with d + 1 columns. First d columns indicate the genetic data (by listing the marker ids) used in the design matrix; last is log likelihood
}
\description{
Perform model fitting for all ordered pairs of markers in a genomic region of interest
}
