% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_pvl.R
\name{scan_pvl}
\alias{scan_pvl}
\title{Perform model fitting for all ordered pairs of markers in a genomic region of interest}
\usage{
scan_pvl(probs, pheno, kinship, covariates = NULL, start_snp1 = 1, n_snp,
  max_iter = 1e+06, max_prec = 1/1e+08)
}
\arguments{
\item{probs}{an array of genotype probabilities for a single chromosome (before eigenrotation)}

\item{pheno}{a matrix of phenotypes (before eigenrotation)}

\item{kinship}{a kinship matrix for one chromosome}

\item{covariates}{a matrix, n subjects by n.cov covariates, where each column is one covariate}

\item{start_snp1}{index of where to start the scan within probs}

\item{n_snp}{the number of (consecutive) snps to include in the scan}

\item{max_iter}{maximum number of iterations for EM algorithm}

\item{max_prec}{stepwise precision for EM algorithm. EM stops once incremental difference in log likelihood is less than max_prec}
}
\value{
a tibble with d + 1 columns. First d columns indicate the genetic data (by listing the marker ids) used in the design matrix; last is log likelihood
}
\description{
Perform model fitting for all ordered pairs of markers in a genomic region of interest
}
\examples{
## define probs
probs_pre <- rbinom(n = 100 * 10, size = 1, prob = 1 / 2)
probs <- array(data = probs_pre, dim = c(100, 1, 10))
dimnames(probs)[[3]] <- paste0("Marker", 1:10)
# define Y
Y_pre <- runif(200)
Y <- matrix(data = Y_pre, nrow = 100)
covariates <- matrix(c(runif(99), NA), nrow = 100, ncol = 1)
cov2 <- matrix(c(covariates[1:99], 10), nrow = 100, ncol = 1)
Y2 <- Y
Y2[1, 2] <- NA
scan_pvl(probs = probs, pheno = Y, kinship = diag(100), start_snp1 = 1, n_snp = 10)
scan_pvl(probs = probs, pheno = Y2, kinship = diag(100), start_snp1 = 1, n_snp = 10)
}
