% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim1.R
\name{sim1}
\alias{sim1}
\title{Simulate a single data set consisting of n subjects and 2 phenotypes for each}
\usage{
sim1(X, B, Vg, Ve, kinship)
}
\arguments{
\item{X}{design matrix (incorporating genotype probabilities from two loci), dn by df}

\item{B}{a matrix of allele effects, f rows by d columns}

\item{Vg}{a genetic covariance matrix, d by d}

\item{Ve}{an error covariance matrix, d by d}

\item{kinship}{a chromosome-specific kinship matrix, n by n}
}
\description{
Simulate a single data set consisting of n subjects and 2 phenotypes for each
}
\examples{
n_mouse <- 20
geno <- rbinom(n = n_mouse, size = 1, prob = 1 / 2)
X <- gemma2::stagger_mats(geno, geno)
B <- matrix(c(1, 2), ncol = 2, nrow = 1)
Vg <- diag(2)
Ve <- diag(2)
sim1(X, B, Vg, Ve, kinship = diag(n_mouse))
}
