% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim1.R
\name{sim1}
\alias{sim1}
\title{Simulate a single data set consisting of n subjects and 2 phenotypes for each}
\usage{
sim1(X, B, Vg, Ve, kinship)
}
\arguments{
\item{X}{design matrix (incorporating genotype probabilities from two loci), dn by df}

\item{B}{a matrix of allele effects, f rows by d columns}

\item{Vg}{a genetic covariance matrix, d by d}

\item{Ve}{an error covariance matrix, d by d}

\item{kinship}{a chromosome-specific kinship matrix, n by n}
}
\description{
Simulate a single data set consisting of n subjects and 2 phenotypes for each
}
