% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_scan_pvl.R
\name{tidy_scan_pvl}
\alias{tidy_scan_pvl}
\title{Tidy the matrix of log likelihood values for further analysis & plotting}
\usage{
tidy_scan_pvl(loglik_mat)
}
\arguments{
\item{loglik_mat}{a (square) matrix of log likelihood values}
}
\description{
Tidy the matrix of log likelihood values for further analysis & plotting
}
