% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_scan_pvl.R
\name{tidy_scan_pvl}
\alias{tidy_scan_pvl}
\title{Tidy the data frame outputted by scan_pvl for further analysis & plotting}
\usage{
tidy_scan_pvl(mytib, pmap)
}
\arguments{
\item{mytib}{outputted dataframe from scan_pvl}

\item{pmap}{physical map (in Mb) for exactly one chromosome, pmap$`5`, for example}
}
\description{
Tidy the data frame outputted by scan_pvl for further analysis & plotting
}
\seealso{
Other profile log-likelihood tibble functions: \code{\link{add_intercepts}},
  \code{\link{add_pmap}},
  \code{\link{assemble_profile_tib}},
  \code{\link{transform_loglik_mat}}
}
\concept{profile log-likelihood tibble functions}
