% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_scan_pvl.R
\name{transform_loglik_mat}
\alias{transform_loglik_mat}
\title{Assemble tibble from matrix of log-likelihood values}
\usage{
transform_loglik_mat(loglik_mat)
}
\arguments{
\item{loglik_mat}{a square matrix of log-likelihood values}
}
\description{
Assemble tibble from matrix of log-likelihood values
}
